+++ addon/query/js-adapter.js
@@ -150,19 +150,68 @@ export default class JSAdapter extends BaseAdapter {
    * @returns {Function}
    */
   buildProjection(query) {
-    if (!query.select || query.select.length === 0) {
+    let select = query.select;
+    let expand = query.expand;
+    if (!select || select.length === 0) {
       return data => data;
     }
 
     return function (data) {
-      return data.map(item => {
+      let dataMap = data.map(item => {
         let r = {};
-        for (let i = 0; i < query.select.length; i++) {
-          r[query.select[i]] = item[query.select[i]];
-        }
 
+        let applySelect = function (r, item, select) {
+          if (!item) {
+            return;
+          }
+
+          let length = select.length;
+          for (let i = 0; i < length; i++) {
+            r[select[i]] = item[select[i]];
+          }
+        };
+
+        applySelect(r, item, select);
+
+        let processExpand = function(r, item, expand) {
+          if (!expand) {
+            return;
+          }
+
+          let expandKeys = Object.keys(expand);
+          let length = expandKeys.length;
+          for (let i = 0; i < length; i++) {
+            let expandKey = expandKeys[i];
+
+            let expandItem = expand[expandKey];
+            let expandItemSelect = expandItem.select;
+
+            if (expandItem.relationship.type === 'belongsTo') {
+              let itemValue = item[expandKey];
+              if (itemValue) {
+                r[expandKey] = {};
+                applySelect(r[expandKey], itemValue, expandItemSelect);
+                processExpand(r[expandKey], itemValue, expandItem.expand);
+              }
+            } else {
+              r[expandKey] = [];
+              let detailsCount = item[expandKey].length;
+              for (let j = 0; j < detailsCount; j++) {
+                let itemValue = item[expandKey][j];
+                if (itemValue) {
+                  r[expandKey].push({});
+                  applySelect(r[expandKey][j], itemValue, expandItemSelect);
+                  processExpand(r[expandKey][j], itemValue, expandItem.expand);
+                }
+              }
+            }
+          }
+        };
+
+        processExpand(r, item, expand);
         return r;
       });
+      return dataMap;
     };
   }