+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Getting of wrong serializer for offline models in online mode.
+- Offline query by projection or select are cut redundand properties for relationships in `js-adapter`.
 
 ### Removed
 - Redundant decoration of online serializer.